Move tests back into `tests` module
